+++ src/components/pages/apps/genes/gsea/bubble/gsea-bubble-dialog.tsx
@@ -148,14 +148,14 @@ export function GseaBubbleDialog({
 
     switch (settings.sortBy) {
       case 'nes':
-        idx = argsort(nes, true)
+        idx = argsort(nes, { reverse: true })
 
         break
       case 'size':
-        idx = argsort(sizes, true)
+        idx = argsort(sizes, { reverse: true })
         break
       case 'pvalue':
-        idx = argsort(log10q, true)
+        idx = argsort(log10q, { reverse: true })
         break
 
       default: