+++ src/components/pages/apps/genes/gsea/gsea-plot/gsea.ts
@@ -1,12 +1,53 @@
 import type { IDBEntity } from '@/interfaces/db-entity'
-import { abs } from '../math/abs'
-import { mean } from '../math/mean'
-import { fisherYatesShuffle } from '../math/random'
+import { abs } from '../../../../../../lib/math/abs'
+import { mean } from '../../../../../../lib/math/mean'
+import { fisherYatesShuffle } from '../../../../../../lib/math/random'
 import { geneSetNames, type IGeneSet, type IRankedGene } from './geneset'
+
+import type { BaseDataFrame } from '@/lib/dataframe/base-dataframe'
+import {
+  getColIdxFromGroup,
+  rowMean,
+  rowStdev,
+} from '@/lib/dataframe/dataframe-utils'
+import { add, sub } from '@/lib/math/add'
+import { argsort } from '@/lib/math/argsort'
+import { div } from '@/lib/math/multiply'
+import type { IClusterGroup } from '../../../../../../lib/cluster-group'
 
 interface EnrichmentResult {
   es: number
   leadingEdge: IRankedGene[]
+}
+
+export function snrRankGenes(
+  df: BaseDataFrame,
+  group1: IClusterGroup,
+  group2: IClusterGroup,
+  epsilon: number = 1e-100
+): IRankedGene[] {
+  // split table for each group
+  const tables = [group1, group2].map((group) => {
+    const colIdx = getColIdxFromGroup(df, group)
+
+    return df.iloc({ cols: colIdx })
+  })
+
+  const names = df.index.strs
+
+  const means = tables.map((df) => rowMean(df))
+
+  const sds = tables.map((df) => rowStdev(df))
+
+  const meanDiffs = sub(means[0]!, means[1]!)
+
+  // add small error so that we never have std 0
+  const sdSum = add(add(sds[0]!, sds[1]!), epsilon)
+  const snr = div(meanDiffs, sdSum)
+
+  const idx = argsort(snr, { reverse: true })
+
+  return idx.map((i, rank) => ({ name: names[i]!, score: snr[i]!, rank }))
 }
 
 function calculateEnrichmentScore(