+++ src/components/pages/apps/matcalc/genesets/geneset-dialog.tsx
@@ -6,9 +6,9 @@ import {
   ActionDialogCardContent,
   ActionDialogRow,
 } from '@/components/dialogs/card/action-dialog-card'
+import type { IGeneSet } from '@/components/pages/apps/genes/gsea/gsea-plot/geneset'
 import { FillButton } from '@/components/plot/fill-dropdown-menu'
 import { DialogTitle } from '@/components/shadcn/ui/themed/v2/dialog'
-import type { IGeneSet } from '@/lib/gsea/geneset'
 import { textToLines } from '@/lib/text/lines'
 import { Textarea } from '@/themed/textarea'
 import { Input } from '@/themed/v2/input'