+++ src/lib/gene/pathway/pathway-loader.ts
@@ -1,6 +1,6 @@
 import { randomHexColor } from '@/lib/color/color'
 
-import type { IGeneSet } from '@/lib/gsea/geneset'
+import type { IGeneSet } from '@/components/pages/apps/genes/gsea/gsea-plot/geneset'
 import { httpFetch } from '@/lib/http/http-fetch'
 import { makeUuid } from '@/lib/id'
 import { textToLines } from '@/lib/text/lines'